# pte-live-api-chat/docs/openapi.yaml
openapi: 3.0.3
info:
  title: pte-live-api-chat
  version: 0.3.0
  description: 聊天与场景消息 API。提供 UserSig 鉴权、单聊/群聊会话、成员、消息、已读、撤回、删除、引用字段、电商直播场景消息，娱乐直播/语聊房房间、成员、麦位、PK、礼物/音效事件，以及 outbox 到 im 的 HTTP/Pulsar 可靠投递 worker。
servers:
  - url: http://127.0.0.1:11504
paths:
...
  /api/v1/im/usersig:
    post:
      summary: 获取腾讯云风格 UserSig 连接凭证
      description: SaaS 商户客户端鉴权接口。服务端按业务 app_id 自动分配或读取 SDKAppID/Secret，返回 user_sig；客户端用 sdkAppID + identifier + userSig 连接 pte-live-im /ws。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                app_id:
                  type: integer
                  example: 10001
                user_id:
                  type: integer
                  example: 10086
                identifier:
                  type: string
                  example: user_10086
                user_type:
                  type: string
                  example: user
                device_id:
                  type: string
                  example: ios-1
                platform:
                  type: string
                  example: app
                scene:
                  type: string
                  example: chat
                expire:
                  type: integer
                  description: 有效期秒，默认 86400，最大 604800。
      responses:
        "200":
          description: UserSig payload
  /api/internal/im/usersig/verify:
    post:
      summary: 内部校验 UserSig
      description: pte-live-im 握手时调用；客户端不直接调用。校验通过后返回业务 app_id、sdk_app_id、identifier、user_id。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sdk_app_id, identifier, user_sig]
              properties:
                sdk_app_id:
                  type: string
                  example: "1400010001"
                identifier:
                  type: string
                  example: user_10086
                user_sig:
                  type: string
      responses:
        "200":
          description: verify result
...